Backend/requirements.txt
Django>=5.0
djangorestframework
django-cors-headers
celery
redis
django-celery-results
pyjwt
requests
dj-database-url
psycopg2-binary
gunicorn
python-dotenv
